Flatten kafka consumer check (#13929)
